libs/ralidator-core/pyproject.toml
[tool.poetry]
name = "ralidator-core"
version = "1.0.0"
description = ""
authors = ["ALeRCE Broker <alercebroker@gmail.com>"]
packages = [{include = "ralidator_core", from = "src"}, {include = "utils", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
PyJWT = "2.3.0"
returns = "0.21.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
